Add report for matched content changes
This adds a report to output content changes by matching subscriber
list over a given period, calculating the number of emails each match
would have generated using a similar technique to the "active_on" scope
on the Subscription model. We want to use this report to see clusters of
changes that contribute to high email volume.