Use ImmediateEmailGenerationService in workers
The change to using this service is to look at resolving an issue where
the ProcessContentChangeAndGenerateEmailsWorker was regularly producing
a deadlock in postgres.

My interpretation of the cause of this is because the content change
worker uses the UnprocessedSubscriptionContentsBySubscriberQuery to pull
out all emails a subscriber needs which could include multiple
subscriptions. Were we to be processing two content changes at the same
time we'd be trying to insert the same unique indexes in separate
processes, thus deadlock.

This change reflects a cut-down in complexity in these workers to
reflect that we can push much of the work to the
ImmediateEmailGenerationService. In theory, if we wanted to speed up
email generation further we could push these out to asynchronous workers
too.

A bit of a bummer of this approach was writing such similar tests to
for each worker. There seems to be enough that is different to justify
them being separate classes, and I don't think we'd benefit from
inheritance, but was a same to write such similar tests.