Add method to bulk populate SubscriptionContent
This method acts as a helper method for the insert_all function allowing
some of the complexities needed for dealing with that to be abstracted
away.

It is named populate_for_content to reflect that there will likely be a
partner method of populate_for_digest_run_subscriber to handle the other
main case for populating this table.