Replace Timecop with ActiveSupport::Testing::TimeHelpers
The ActiveSupport helpers are nicer to use because they automatically
round times to the second resolving the frequent pain-point of Postgres
and Ruby disagreeing on the precision of a time. This also saves a
dependency.

I've done all the simple tests to swap (and my what a variety of ways
has it been called). I've got two feature tests to fix too that are a
bit more complex.