Extend stats for immediate content changes
Previously we could only see the realtime number of emails being
processed by the system, noting that almost all email we send is
due to a content change. This changes the processing of immediate
emails for a content change to send a breakdown of the generated
email by document type and publishing app, to improve visibility
of the effect of publishing on the volume of email we send, and to
help reassure ourselves about its source.

Note that this metric is not applicable for digest emails, because
of the many-to-one link between content changes and the email that
is sent. For digest emails, the number of subscriptions is a better
(maximum) indicator of the volume of email we expect to send.