Prevent Sentry alerts for PG::Errors during data sync
This ports over code from Content Publisher [1] that prevents Sentry
from being notified of PostgreSQL errors during a time period in
integration and staging environment.

Looking in Sentry there is _a lot_ of errors sent during the data sync.
In integration there were ~56k errors during the last 14 days and in
staging there were in the region of ~42k.

Arguably, this reveals a bunch of flaws in our data sync process as it
looks like we produce errors for a number of hours each night.

The tricky thing in this PR is deciding on what is a reasonable time to
start reporting errors again, as outside of data syncs we do want to
be informed of DB errors in integration and staging. Looking at recent
data I can see the data syncs occur until around 6am, so I felt 7am
should provide sufficient buffer and be unlikely to intrude on many
peoples working days. However this may need tweaking if we find
ourselves alerted.

[1]: https://github.com/alphagov/content-publisher/blob/2406ab6fc7bf5bdb2bd7d05c7f585b745c2ad061/config/initializers/govuk_error.rb