Stop exceptions during Subscriber find or create flow
If more than one concurrent request is sent to Email Alert API to create
a subscription a 500 response can be returned. This can be caused by the
find and create block failing to find a user in multiple requests and
only the first request will be able to create the user.

The approach taken to resolve this is to use a retryable transaction to
find and create the user. The expectation is that if the initial find
and create fails running the code again will succeed because there is
now a record to find on retry.

We use a near identical approach in Publishing API [1] however this one
has been tailored to this individual model as this has a complication in
the address needing to be looked up by lowercase and wanting additional
data in the create.

The impetus for this was seeing Sentry errors come through for Email
Alert Frontend receiving a 500 response from Email Alert API which was
due to a ActiveRecord::RecordNotUnique

I can see that this was caused by a user clicking to activate two
different subscriptions at the same time.

[1]: https://github.com/alphagov/publishing-api/blob/427814de539d804a2c5fcd491181b9ee997eae4c/app/models/concerns/find_or_create_locked.rb#L5-L19