Mark email as sent on successful request
We want to mark an email as sent and record the time it was sent at when
we send a successful email request to Notify.

We also no longer want to be updating the email models status after the
initial delivery to Notify. An example of this is where we're setting
the status of an email to failed following a status update. This is
confusing if an email's status changes from `sent` to `failed`. The
delivery attempt will show if the email has actually been delivered or
not.