Allow all emails to be archivable and deleteable
This change allows all emails to be archivable which allows them also to
be deleted.

Previously emails were only archived if they did not have a status of
pending. This commit changes that.

This is part of the cleanup work so we don't end up with lingering
emails and subscriptions contents in the system which will never be
deleted or recovered.

This commit also removes the scopes of 'deleteable' and 'archiveable'
as they each only had one filter so seemed unecessary to keep them.

Trello:
https://trello.com/c/tfu9zXuy/534-delete-unused-data-after-a-year