Replace descriptions with hard-coded Brexit URL
Previously we added a list description field to support a custom
link for travel advice lists [1], which was subsequently removed
[2]. Only the Brexit Checker now uses descriptions, to provide a
CTA URL to the results page for that list [3].

This replaces the generic description field with a presenter that
hard-codes the same URL text used across the ~70K lists associated
with the Brexit Checker. Using hard-coded text seems like a better
compromise than a list description since:

- It's easier to update (vs. a large DB migration).
- It's purpose is clearer (it was never really a description).
- It's more DRY (by reusing the URL field on a list).

The functionality remains fairly generic, and could be extended to
show a generic URL for other lists. It has the added advantage that
its single use case is now explicit in code, so we can also easily
remove it if it becomes unused.

[1]: https://github.com/alphagov/email-alert-api/pull/1006
[2]: https://github.com/alphagov/email-alert-api/pull/1135
[3]: https://github.com/alphagov/finder-frontend/blob/2a07fa1bcfd68ce640225f3646aeb8d81b3a823a/app/controllers/brexit_checker_controller.rb#L186