Update digest email template to match new design
This follows a similar approach to I18n keys [1] and testing [2],
as for individual-type emails.

Note that, unlike individual emails, the feature specs are intent
on testing the full content of the emails generated, which isn't
so easy now they may contain randomised one-click links. For now,
this switches to using an "include" matcher and avoids matching
the footer part of the email, which is covered by builder specs.
We should explore how to DRY up these specs in future work.

[1]: https://github.com/alphagov/email-alert-api/pull/1520
[2]: https://github.com/alphagov/email-alert-api/blob/693babb635c58049d9633dd9123ece4670748ba0/spec/builders/immediate_email_builder_spec.rb