Add standard footer to bulk emails
This changes the bulk email builder to pass around subscriptions for
each subscriber, so that we can use the latest of these in generating
a standard footer for each email, with a one-click unsubscribe link.

Although the subscription to choose when de-duping is arbitrary at
this point, it does generally make sense to choose the latest one, and
this will become even more necessary in a later commit.